refactor: normalize parameter order and timeout types
AssertUpdateSecret now puts namespace and clusterName right after env,
matching the (env, namespace, cluster, ...) convention used by every
other helper. AssertFastFailOver groups quickDeletionPeriod with the
other timing knobs at the end of the signature.

Three replication helpers and AssertFastFailOver switch the timeout
parameters from int32 to int so they line up with testTimeouts map
values and the rest of the asserts API; the test files' variable
declarations follow.

+++ tests/e2e/backup_restore_minio_test.go
@@ -673,7 +673,7 @@ var _ = Describe("MinIO - Backup and restore", Label(tests.LabelBackupRestore), 
 					env,
 					namespace,
 					firstClusterName,
-					int32(testTimeouts[timeouts.ClusterIsReadyQuick]),
+					testTimeouts[timeouts.ClusterIsReadyQuick],
 				)
 			})
